+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.4] - 2021-11-03
+
 ### Fixed
 - Adjustment to OSGB36 datum transform e.g. EPSG: 27700
 - GeocentricConverter equality check after grid shift WGS param override e.g. EPSG: 27700 [#32]
@@ -61,7 +63,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix possible `null` dereference [#16](https://github.com/locationtech/proj4j/pull/16)
 - Fix `cea` (Cylindrical Equal Area) projection [#10](https://github.com/locationtech/proj4j/pull/10)
 
-[Unreleased]: https://github.com/locationtech/proj4j/compare/v1.1.3...HEAD
+[Unreleased]: https://github.com/locationtech/proj4j/compare/v1.1.4...HEAD
+[1.1.4]: https://github.com/locationtech/proj4j/compare/v1.1.3...v1.1.4
 [1.1.3]: https://github.com/locationtech/proj4j/compare/v1.1.2...v1.1.3
 [1.1.2]: https://github.com/locationtech/proj4j/compare/v1.1.1...v1.1.2
 [1.1.1]: https://github.com/locationtech/proj4j/compare/v1.1.0...v1.1.1